Run larger transfers as background jobs :sparkles:
Before this change, the API request from the client could time out if it
has to wait too long for a transfer to complete in the foreground.
Blocking the request also locks up a PHP-FPM process which should be
responding to other requests.

Additionally, by sending larger transfers to the background, the user
can continue working while the task is running, rather than having to
wait at a loading screen for an unknown length of time.

TODO: Send a notification when a transfer finishes.